refactor: Clean up code style and fix multiple minor issues
This commit includes multiple code optimizations and fixes:
1. Add multiple clippy lint allow annotations to resolve code warnings
2. Replace legacy syntax with more concise Rust 2018+ features, such as using `is_empty()` instead of `len() > 0`, using `if let` chained syntax to simplify checks
3. Fix some type conversion and calculation logic issues, including using `div_ceil` instead of manual byte count calculation
4. Refactor some conditional logic, use match syntax instead of multi-layer if nesting
5. Adjust comment formatting and code indentation to improve readability
6. Optimize FTP/SFTP/BT related address parsing logic, support IPv6 bracket format
7. Fix some test cases and documentation comments
8. Adjust Cargo dependency configuration, move `tokio-uring` to Linux platform-specific dependencies
9. Refactor HTTP Content-Range parsing logic, compatible with more server return formats
10. Supplement some missing method implementations and test cases

+++ aria2-core/src/dht/routing_table_ser.rs
@@ -460,7 +460,7 @@ mod tests {
         for node in &result.nodes {
             assert_eq!(
                 node.id().as_bytes()[0],
-                ((node.addr().port() as u16 - 6881u16) & 0xFF) as u8
+                ((node.addr().port() - 6881u16) & 0xFF) as u8
             );
         }
     }
